clean up FilmedApi, get rid of a lot of cruft
